+++ Dockerfile
@@ -9,6 +9,11 @@ WORKDIR /home/ecscontaineruser
 # Copy source file to the image work directory
 COPY . .
 
+# Update packages
+RUN apt-get update && \
+    apt-get upgrade -y && \
+    pip install --upgrade pip
+
 # Install the requirements
 RUN pip install -r requirements.txt
 
@@ -19,4 +24,5 @@ USER ecscontaineruser
 EXPOSE 3000
 
 # Entrypoint
-ENTRYPOINT [ "python", "-m", "src.app" ]
+# ENTRYPOINT [ "python", "-m", "src.app" ]
+ENTRYPOINT ["gunicorn", "-w", "4", "-b", "0.0.0.0:3000", "src.app:app"]
