+++ release.version
@@ -1,1 +1,1 @@
-v1.0.4
+v1.0.5
