+++ requirements.txt
@@ -1,2 +1,3 @@
 flask==3.0.0
 boto3>=1.30.0
+gunicorn>=21.1.0
