+++ src/controllers/__init__.py
@@ -1,0 +1,2 @@
+# /app/src/controllers/__init__.py
+from .prayer_controller import prayer_bp
